cleaned up connectors and infoCard
