+++ frontend/control-center/src/pages/Connectors/ConnectorConfig/index.tsx
@@ -17,7 +17,7 @@ import {
 import {LinkButton, InfoButton} from 'components';
 import {NotificationModel, Source, ComponentInfo} from 'model';
 import {ReactComponent as ArrowLeftIcon} from 'assets/images/icons/leftArrowCircle.svg';
-import {ConnectNewDialogflow} from '../Providers/Dialogflow/ConnectNewDialogflow';
+import {DialogflowConnect} from '../Providers/Dialogflow/DialogflowConnect';
 import {ConnectNewZendesk} from '../Providers/Zendesk/ConnectNewZendesk';
 import {ConnectNewSalesforce} from '../Providers/Salesforce/ConnectNewSalesforce';
 import {ConfigStatusButton} from '../ConfigStatusButton';
@@ -77,14 +77,13 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
   } = props;
 
   const {channelId, source} = useParams();
-  const connectorConfiguration = useSelector((state: StateModel) => state.data.connector);
-
+  const connectors = useSelector((state: StateModel) => state.data.connector);
   const [connectorInfo, setConnectorInfo] = useState<ComponentInfo | null>(null);
-  const configKey = connectorInfo && removePrefix(connectorInfo?.name);
+  const componentName = connectorInfo && removePrefix(connectorInfo?.name);
   const [currentPage] = useState(Pages.createUpdate);
   const [configurationModal, setConfigurationModal] = useState(false);
   const [notification, setNotification] = useState<NotificationModel>(null);
-  const [isEnabled, setIsEnabled] = useState<boolean | null>(components[connectorInfo && configKey]?.enabled);
+  const [isEnabled, setIsEnabled] = useState<boolean | null>(components[connectorInfo && componentName]?.enabled);
   const [isPending, setIsPending] = useState(false);
   const [isConfigured, setIsConfigured] = useState(false);
   const [backTitle, setBackTitle] = useState('Connectors');
@@ -103,15 +102,12 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
   }, []);
 
   useEffect(() => {
-    if (connectorInfo && connectorConfiguration && connectorConfiguration[connectorInfo.name]) {
-      if (
-        Object.entries(connectorConfiguration[connectorInfo.name]) &&
-        Object.entries(connectorConfiguration[connectorInfo.name]).length > 0
-      ) {
+    if (connectorInfo && connectors && connectors[componentName]) {
+      if (Object.keys(connectors[componentName]).length > 0) {
         setIsConfigured(true);
       }
     }
-  }, [connectorInfo, connectorConfiguration]);
+  }, [connectorInfo, connectors]);
 
   useEffect(() => {
     getConnectorsConfiguration().catch((error: Error) => {
@@ -151,7 +147,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
 
   useEffect(() => {
     if (config && connectorInfo) {
-      setIsEnabled(config?.components[configKey]?.enabled);
+      setIsEnabled(config?.components[componentName]?.enabled);
     }
   }, [config, connectorInfo, components]);
 
@@ -173,7 +169,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
       payload = {
         components: [
           {
-            name: connectorInfo && connectorInfo.name,
+            name: connectorInfo && removePrefix(connectorInfo.name),
             enabled: true,
             data: {
               projectId: projectId,
@@ -195,7 +191,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
       payload = {
         components: [
           {
-            name: connectorInfo && connectorInfo.name,
+            name: connectorInfo && removePrefix(connectorInfo.name),
             enabled: true,
             data: {
               domain: domain,
@@ -213,7 +209,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
       payload = {
         components: [
           {
-            name: connectorInfo && connectorInfo.name,
+            name: connectorInfo && removePrefix(connectorInfo.name),
             enabled: true,
             data: {
               url: url,
@@ -232,7 +228,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
       payload = {
         components: [
           {
-            name: connectorInfo && connectorInfo?.name,
+            name: connectorInfo && removePrefix(connectorInfo.name),
             enabled: true,
             data: {
               webhookUrl: webhookUrl,
@@ -250,7 +246,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
       payload = {
         components: [
           {
-            name: connectorInfo && connectorInfo?.name,
+            name: connectorInfo && removePrefix(connectorInfo.name),
             enabled: true,
             data: {
               appId: appId,
@@ -281,7 +277,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
   const PageContent = () => {
     if (connector === Source.dialogflow) {
       return (
-        <ConnectNewDialogflow
+        <DialogflowConnect
           createNewConnection={createNewConnection}
           isEnabled={isEnabled}
           isConfigured={isConfigured}
@@ -359,7 +355,7 @@ const ConnectorConfig = (props: ConnectorConfigProps) => {
   const enableDisableComponentToggle = () => {
     setConfigurationModal(false);
     setIsPending(true);
-    enableDisableComponent({components: [{name: configKey, enabled: !isEnabled}]})
+    enableDisableComponent({components: [{name: componentName, enabled: !isEnabled}]})
       .then(() => {
         setNotification({
           show: true,
