+++ frontend/control-center/src/pages/Connectors/InfoCard/index.module.scss
@@ -32,14 +32,6 @@
   }
 }
 
-.installed {
-  height: 98px;
-}
-
-.notInstalled {
-  height: 118px;
-}
-
 .channelLogoInstagram {
   margin-right: 10px;
   height: 40px;
