+++ frontend/control-center/src/pages/Connectors/Providers/Dialogflow/ConnectorDialogflow.tsx
@@ -1,0 +1,9 @@
+import {Source} from 'model';
+import React from 'react';
+import ConnectorConfig from '../../ConnectorConfig';
+
+const ConnectorDialogflow = () => {
+  return <ConnectorConfig connector={Source.dialogflow} />;
+};
+
+export default ConnectorDialogflow;
